+++ ci4/app/Interfaces/BoolInterface.php
@@ -1,0 +1,7 @@
+<?php namespace App\Interfaces;
+/**
+ * @property bool $value
+ */
+interface BoolInterface {
+
+}